Cover Responses stream error events in CI path
